% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixff-package.R
\docType{package}
\name{mixff}
\alias{mixff-package}
\alias{mixff}
\title{mixff: A package}
\description{
The mixff package provides ...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
