% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcfa-package.R
\docType{package}
\name{mmcfa}
\alias{mmcfa-package}
\alias{mmcfa}
\title{mmcfa: A package}
\description{
The mmcfa package provides ...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
